update
opedirの関数をprocessing に統合
mainでの実装時にうまくいかなかったため
